Moved examples folder into package
