+++ docs/tutorials/minimal_example.rst
@@ -17,7 +17,7 @@ If you are not already familiar with OpenMDAO, we strongly recommend going throu
 Imports
 =======
 
-.. literalinclude:: ../../examples/minimal.py
+.. literalinclude:: ../../openconcept/examples/minimal.py
     :start-after: # rst Imports (beg)
     :end-before: # rst Imports (end)
 
@@ -42,14 +42,14 @@ Drag is computed as lift divided by lift-to-drag ratio.
 Let's first take a look at the code for the whole aircraft model, and then we'll explain each part.
 The whole model looks like this:
 
-.. literalinclude:: ../../examples/minimal.py
+.. literalinclude:: ../../openconcept/examples/minimal.py
     :start-after: # rst Aircraft (beg)
     :end-before: # rst Aircraft (end)
 
 Options
 -------
 
-.. literalinclude:: ../../examples/minimal.py
+.. literalinclude:: ../../openconcept/examples/minimal.py
     :start-after: # rst Options
     :end-before: # rst Setup
     :dedent: 4
@@ -63,7 +63,7 @@ These two options are **required** for all OpenConcept aircraft models:
 Setup
 -----
 
-.. literalinclude:: ../../examples/minimal.py
+.. literalinclude:: ../../openconcept/examples/minimal.py
     :start-after: # rst Setup
     :end-before: # rst Compute
     :dedent: 4
@@ -150,7 +150,7 @@ In practice, analytically defining the partial derivatives offers more accurate 
 Compute
 -------
 
-.. literalinclude:: ../../examples/minimal.py
+.. literalinclude:: ../../openconcept/examples/minimal.py
     :start-after: # rst Compute
     :end-before: # rst Aircraft (end)
     :dedent: 4
@@ -163,7 +163,7 @@ The drag is equal to the lift divided by L/D, where lift is computed using the d
 Mission
 =======
 
-.. literalinclude:: ../../examples/minimal.py
+.. literalinclude:: ../../openconcept/examples/minimal.py
     :start-after: # rst Mission (beg)
     :end-before: # rst Mission (end)
 
@@ -201,7 +201,7 @@ Setup problem
 Now that we have the necessary models.
 The last step before running the model is setting up the OpenMDAO problem and providing the necessary values to define the mission profile.
 
-.. literalinclude:: ../../examples/minimal.py
+.. literalinclude:: ../../openconcept/examples/minimal.py
     :start-after: # rst Setup problem (beg)
     :end-before: # rst Setup problem (end)
 
@@ -221,7 +221,7 @@ Run it!
 
 Finally, we actually run the analysis.
 
-.. literalinclude:: ../../examples/minimal.py
+.. literalinclude:: ../../openconcept/examples/minimal.py
     :start-after: # rst Run (beg)
     :end-before: # rst Run (end)
 
@@ -238,7 +238,7 @@ The plot it generates should look like this:
 The N2 diagram for the model is the following:
 
 .. embed-n2::
-  ../examples/minimal.py
+  ../openconcept/examples/minimal.py
 
 Summary
 =======
@@ -253,4 +253,4 @@ In the next tutorial, we'll develop a more comprehensive aircraft model that is 
 
 The final script looks like this:
 
-.. literalinclude:: ../../examples/minimal.py
+.. literalinclude:: ../../openconcept/examples/minimal.py
