+++ openconcept/examples/B738_Dymos.py
@@ -1,14 +1,11 @@
 from __future__ import division
-import sys
-import os
 import numpy as np
 
-sys.path.insert(0, os.getcwd())
 import openmdao.api as om
 import openconcept.api as oc
 # imports for the airplane model itself
 from openconcept.analysis.aerodynamics import PolarDrag
-from examples.aircraft_data.B738 import data as acdata
+from openconcept.examples.aircraft_data.B738 import data as acdata
 from openconcept.analysis.performance.mission_profiles import MissionWithReserve
 from openconcept.components.cfm56 import CFM56
