+++ openconcept/examples/HybridTwin.py
@@ -1,9 +1,8 @@
 from __future__ import division
-import sys
 import os
 import logging
 import numpy as np
-sys.path.insert(0, os.getcwd())
+
 from openmdao.api import Problem, Group, ScipyOptimizeDriver
 from openmdao.api import BalanceComp, ExplicitComponent, ExecComp, SqliteRecorder
 from openmdao.api import DirectSolver, IndepVarComp, NewtonSolver, BoundsEnforceLS
@@ -11,24 +10,16 @@ from openmdao.api import DirectSolver, IndepVarComp, NewtonSolver, BoundsEnforce
 # imports for the airplane model itself
 from openconcept.analysis.aerodynamics import PolarDrag
 from openconcept.utilities.math import AddSubtractComp
-from openconcept.utilities.math.max_min_comp import MaxComp
 from openconcept.utilities.math.integrals import Integrator
 from openconcept.utilities.dvlabel import DVLabel
-from examples.methods.weights_twin_hybrid import TwinSeriesHybridEmptyWeight
-from examples.propulsion_layouts.thermal_series_hybrid import TwinSeriesHybridElectricPropulsionRefrigerated
-from examples.methods.costs_commuter import OperatingCost
+from openconcept.examples.methods.weights_twin_hybrid import TwinSeriesHybridEmptyWeight
+from openconcept.examples.propulsion_layouts.simple_series_hybrid import TwinSeriesHybridElectricPropulsionSystem
+from openconcept.examples.methods.costs_commuter import OperatingCost
 from openconcept.utilities.dict_indepvarcomp import DictIndepVarComp
-from examples.aircraft_data.KingAirC90GT import data as acdata
-from openconcept.analysis.performance.mission_profiles import BasicMission
+from openconcept.examples.aircraft_data.KingAirC90GT import data as acdata
+from openconcept.analysis.performance.mission_profiles import FullMissionAnalysis
 from openconcept.utilities.linearinterp import LinearInterpolator
 from openconcept.utilities.visualization import plot_trajectory
-
-"""
-WARNING: This example has known convergence problems because of the chiller in the
-         propulsion layout (TwinSeriesHybridElectricPropulsionRefrigerated)
-
-Eytan Adler, 27/10/2021
-"""
 
 class AugmentedFBObjective(ExplicitComponent):
     def setup(self):
@@ -42,8 +33,8 @@ class AugmentedFBObjective(ExplicitComponent):
 
 class SeriesHybridTwinModel(Group):
     """
-    A custom model specific to a series hybrid twin turboprop-class airplane with an active
-    thermal management system. This class will be passed in to the mission analysis code.
+    A custom model specific to a series hybrid twin turboprop-class airplane
+    This class will be passed in to the mission analysis code.
 
     """
     def initialize(self):
@@ -57,7 +48,6 @@ class SeriesHybridTwinModel(Group):
         # any control variables other than throttle and braking need to be defined here
         controls = self.add_subsystem('controls', IndepVarComp(), promotes_outputs=['*'])
         controls.add_output('proprpm',val=np.ones((nn,))*2000, units='rpm')
-        controls.add_output('ac|propulsion|thermal|hx|mdot_coolant', val=0.1*np.ones((nn,)), units='kg/s')
 
         # assume TO happens on battery backup
         if flight_phase in ['climb', 'cruise','descent']:
@@ -74,7 +64,7 @@ class SeriesHybridTwinModel(Group):
                                       "ac|weights*", 'duration']
 
         self.add_subsystem('propmodel',
-                           TwinSeriesHybridElectricPropulsionRefrigerated(num_nodes=nn),
+                           TwinSeriesHybridElectricPropulsionSystem(num_nodes=nn),
                            promotes_inputs=propulsion_promotes_inputs,
                            promotes_outputs=propulsion_promotes_outputs)
         self.connect('proprpm', ['propmodel.prop1.rpm', 'propmodel.prop2.rpm'])
@@ -87,28 +77,16 @@ class SeriesHybridTwinModel(Group):
             cd0_source = 'ac|aero|polar|CD0_TO'
         self.add_subsystem('drag', PolarDrag(num_nodes=nn),
                            promotes_inputs=['fltcond|CL', 'ac|geom|*', ('CD0', cd0_source),
-                                            'fltcond|q', ('e', 'ac|aero|polar|e')])
+                                            'fltcond|q', ('e', 'ac|aero|polar|e')],
+                           promotes_outputs=['drag'])
 
         self.add_subsystem('OEW',TwinSeriesHybridEmptyWeight(),
-                           promotes_inputs=['*', ('P_TO','ac|propulsion|engine|rating')])
+                           promotes_inputs=['*',('P_TO','ac|propulsion|engine|rating')],
+                           promotes_outputs=['OEW'])
         self.connect('propmodel.propellers_weight', 'W_propeller')
         self.connect('propmodel.eng1.component_weight', 'W_engine')
         self.connect('propmodel.gen1.component_weight', 'W_generator')
         self.connect('propmodel.motors_weight', 'W_motors')
-
-        hxadder = AddSubtractComp()
-        hxadder.add_equation('OEW',['OEW_orig','W_hx','W_coolant'],scaling_factors=[1,1,1],units='kg')
-        hxadder.add_equation('drag',['drag_orig','drag_hx'], vec_size=nn, units='N', scaling_factors=[1,1])
-        hxadder.add_equation('area_constraint',['hx_frontal_area','nozzle_area'],units='m**2',scaling_factors=[1,-1])
-        self.add_subsystem('hxadder',hxadder, promotes_inputs=[('W_coolant','ac|propulsion|thermal|hx|coolant_mass')],promotes_outputs=['OEW','drag'])
-        self.connect('drag.drag','hxadder.drag_orig')
-        self.connect('OEW.OEW','hxadder.OEW_orig')
-        self.connect('propmodel.hx.component_weight','hxadder.W_hx')
-        self.connect('propmodel.duct.drag','hxadder.drag_hx')
-        self.connect('propmodel.hx.frontal_area','hxadder.hx_frontal_area')
-        self.add_subsystem('nozzle_area', MaxComp(num_nodes=nn, units='m**2'))
-        self.connect('propmodel.area_nozzle', 'nozzle_area.array')
-        self.connect('nozzle_area.max','hxadder.nozzle_area')
         intfuel = self.add_subsystem('intfuel', Integrator(num_nodes=nn, method='simpson', diff_units='s',
                                                               time_setup='duration'), promotes_inputs=['*'], promotes_outputs=['*'])
         intfuel.add_integrand('fuel_used', rate_name='fuel_flow', val=1.0, units='kg')
@@ -158,13 +136,6 @@ class ElectricTwinAnalysisGroup(Group):
         dv_comp.add_output_from_dict('ac|propulsion|propeller|diameter')
         dv_comp.add_output_from_dict('ac|propulsion|generator|rating')
         dv_comp.add_output_from_dict('ac|propulsion|motor|rating')
-
-        dv_comp.add_output('ac|propulsion|thermal|hx|coolant_mass',val=10.,units='kg')
-        dv_comp.add_output('ac|propulsion|thermal|hx|channel_width',val=1.,units='mm')
-        dv_comp.add_output('ac|propulsion|thermal|hx|channel_height',20.,units='mm')
-        dv_comp.add_output('ac|propulsion|thermal|hx|channel_length',val=0.2,units='m')
-        dv_comp.add_output('ac|propulsion|thermal|hx|n_parallel',val=50,units=None)
-        dv_comp.add_output('ac|propulsion|thermal|hx|n_wide_cold',val=430,units=None)
         dv_comp.add_output('ac|propulsion|battery|specific_energy',val=300,units='W*h/kg')
 
         dv_comp.add_output_from_dict('ac|num_passengers_max')
@@ -173,14 +144,10 @@ class ElectricTwinAnalysisGroup(Group):
 
         mission_data_comp = self.add_subsystem('mission_data_comp',IndepVarComp(),promotes_outputs=["*"])
         mission_data_comp.add_output('batt_soc_target', val=0.1, units=None)
-        mission_data_comp.add_output('T_motor_initial', val=15, units='degC')
-        mission_data_comp.add_output('T_batt_initial', val=10.1, units='degC')
 
-        # Ensure that any state variables are connected across the mission as intended
-        analysis = self.add_subsystem('analysis',BasicMission(num_nodes=nn,
+        analysis = self.add_subsystem('analysis',FullMissionAnalysis(num_nodes=nn,
                                                                      aircraft_model=SeriesHybridTwinModel),
-                                                 promotes_inputs=['*'],promotes_outputs=
-                                                 ['*'])
+                                                 promotes_inputs=['*'],promotes_outputs=['*'])
 
         margins = self.add_subsystem('margins',ExecComp('MTOW_margin = MTOW - OEW - total_fuel - W_battery - payload',
                                                         MTOW_margin={'units':'lbm','val':100},
@@ -199,25 +166,20 @@ class ElectricTwinAnalysisGroup(Group):
         self.connect('ac|weights|MTOW','aug_obj.ac|weights|MTOW')
         self.connect('descent.fuel_used_final','aug_obj.fuel_burn')
 
-        self.connect('T_motor_initial','climb.propmodel.motorheatsink.T_initial')
-        self.connect('T_batt_initial','climb.propmodel.batteryheatsink.T_initial')
-
 def configure_problem():
     prob = Problem()
     prob.model= ElectricTwinAnalysisGroup()
-    prob.model.nonlinear_solver=NewtonSolver(iprint=2)
+    prob.model.nonlinear_solver=NewtonSolver(iprint=1)
     prob.model.options['assembled_jac_type'] = 'csc'
     prob.model.linear_solver = DirectSolver(assemble_jac=True)
     prob.model.nonlinear_solver.options['solve_subsystems'] = True
     prob.model.nonlinear_solver.options['maxiter'] = 10
-    prob.model.nonlinear_solver.options['atol'] = 1e-8
-    prob.model.nonlinear_solver.options['rtol'] = 1e-8
-    prob.model.nonlinear_solver.linesearch = BoundsEnforceLS()
-    # prob.model.nonlinear_solver.linesearch.options['print_bound_enforce'] = True
+    prob.model.nonlinear_solver.options['atol'] = 1e-7
+    prob.model.nonlinear_solver.options['rtol'] = 1e-7
     return prob
 
-def set_values(prob, num_nodes, design_range, spec_energy):
-    # set some (required) mission parameters. Each phase needs a vertical and air-speed
+def set_values(prob, num_nodes, design_range, spec_energy): 
+    # set some (required) mission parameters. Each pahse needs a vertical and air-speed
     # the entire mission needs a cruise altitude and range
     prob.set_val('climb.fltcond|vs', np.ones((num_nodes,))*1500, units='ft/min')
     prob.set_val('climb.fltcond|Ueas', np.ones((num_nodes,))*124, units='kn')
@@ -230,22 +192,21 @@ def set_values(prob, num_nodes, design_range, spec_energy):
     prob.set_val('mission_range',design_range,units='NM')
     prob.set_val('payload',1000,units='lb')
     prob.set_val('ac|propulsion|battery|specific_energy', spec_energy, units='W*h/kg')
+
+    # (optional) guesses for takeoff speeds may help with convergence
+    prob.set_val('v0v1.fltcond|Utrue',np.ones((num_nodes))*50,units='kn')
+    prob.set_val('v1vr.fltcond|Utrue',np.ones((num_nodes))*85,units='kn')
+    prob.set_val('v1v0.fltcond|Utrue',np.ones((num_nodes))*85,units='kn')
+
     # set some airplane-specific values
     prob['analysis.cruise.acmodel.OEW.const.structural_fudge'] = 2.0
     prob['ac|propulsion|propeller|diameter'] = 2.2
     prob['ac|propulsion|engine|rating'] = 1117.2
 
-    # Turn off the refrigerator during certain segments
-    prob['analysis.cruise.acmodel.propmodel.refrig.control.bypass_start'] = 1
-    prob['analysis.cruise.acmodel.propmodel.refrig.control.bypass_end'] = 1
-
-    # set the initial battery SOC to match the HybridTwin_thermal after takeoff
-    prob.set_val('climb.propmodel.batt1.SOC_initial', 0.8611499461827815, units=None)
-
-def run_hybrid_twin_active_thermal_analysis(plots=False):
+def run_hybrid_twin_analysis(plots=False):
     prob = configure_problem()
     prob.setup(check=False)
-    prob['cruise.hybridization'] = 0.05778372636876463
+    prob['cruise.hybridization'] = 0.05840626452293813
     set_values(prob, 11, 500, 450)
     prob.run_model()
     if plots:
@@ -255,45 +216,42 @@ def run_hybrid_twin_active_thermal_analysis(plots=False):
 def show_outputs(prob):
     # print some outputs
     vars_list = ['ac|weights|MTOW','climb.OEW','descent.fuel_used_final',
-                 'descent.propmodel.batt1.SOC_final','cruise.hybridization',
+                 'rotate.range_final','descent.propmodel.batt1.SOC_final','cruise.hybridization',
                  'ac|weights|W_battery','margins.MTOW_margin',
                  'ac|propulsion|motor|rating','ac|propulsion|generator|rating','ac|propulsion|engine|rating',
-                 'ac|geom|wing|S_ref',
-                 'cruise.propmodel.duct.drag', 'ac|propulsion|thermal|hx|coolant_mass',
-                 'climb.propmodel.duct.mdot']
+                 'ac|geom|wing|S_ref','v0v1.Vstall_eas','v0v1.takeoff|vr',
+                 'engineoutclimb.gamma']
     units = ['lb','lb','lb',
-             None, None,
+             'ft', None, None,
              'lb','lb',
              'hp','hp','hp',
-             'ft**2',
-             'lbf','lb',
-             'lb/s']
+             'ft**2','kn','kn',
+             'deg']
     nice_print_names = ['MTOW', 'OEW', 'Fuel used',
-                        'Final state of charge', 'Cruise hybridization',
+                        'TOFL (over 35ft obstacle)', 'Final state of charge', 'Cruise hybridization',
                         'Battery weight','MTOW margin',
                         'Motor rating', 'Generator rating', 'Engine rating',
-                        'Wing area',
-                        'Coolant duct cruise drag', 'Coolant mass',
-                        'Coolant duct mass flow']
+                        'Wing area', 'Stall speed', 'Rotate speed',
+                        'Engine out climb angle']
     print("=======================================================================")
     for i, thing in enumerate(vars_list):
         print(nice_print_names[i]+': '+str(prob.get_val(thing,units=units[i])[0])+' '+str(units[i]))
-
-    print('Motor temps (climb): '+str(prob.get_val('climb.propmodel.motorheatsink.T',units='degC')))
-    print('Battery temps (climb): '+str(prob.get_val('climb.propmodel.batteryheatsink.T',units='degC')))
 
     # plot some stuff
     plots = True
     if plots:
         x_var = 'range'
         x_unit = 'NM'
-        y_vars = ['fltcond|h','fltcond|Ueas','fuel_used','throttle','fltcond|vs','propmodel.batt1.SOC',
-                  'propmodel.motorheatsink.T','propmodel.batteryheatsink.T']
-        y_units = ['ft','kn','lbm',None,'ft/min',None,'degC','degC','inch**2']
+        y_vars = ['fltcond|h','fltcond|Ueas','fuel_used','throttle','fltcond|vs','propmodel.batt1.SOC']
+        y_units = ['ft','kn','lbm',None,'ft/min',None]
         x_label = 'Range (nmi)'
-        y_labels = ['Altitude (ft)', 'Veas airspeed (knots)', 'Fuel used (lb)', 'Throttle setting', 'Vertical speed (ft/min)', 'Battery SOC', 'Motor temp', 'Battery temp', 'HX area']
+        y_labels = ['Altitude (ft)', 'Veas airspeed (knots)', 'Fuel used (lb)', 'Throttle setting', 'Vertical speed (ft/min)', 'Battery SOC']
+        phases = ['v0v1','v1vr','v1v0','rotate']
+        plot_trajectory(prob, x_var, x_unit, y_vars, y_units, phases,
+                        x_label=x_label, y_labels=y_labels, marker='-',
+                        plot_title='Takeoff Profile')
 
-        phases = ['climb','cruise','descent']
+        phases = ['v0v1','v1vr','v1v0','rotate','climb','cruise','descent']
         plot_trajectory(prob, x_var, x_unit, y_vars, y_units, phases,
                         x_label=x_label, y_labels=y_labels, marker='-',
                         plot_title='Full Mission Profile')
@@ -305,25 +263,23 @@ if __name__ == "__main__":
 
     if run_type == 'example':
         # runs a default analysis-only mission (no optimization)
-        run_hybrid_twin_active_thermal_analysis(plots=True)
+        run_hybrid_twin_analysis(plots=True)
 
     else:
         # can run a sweep of design range and spec energy (not tested)
         #design_ranges = [300,350,400,450,500,550,600,650,700]
         #specific_energies = [250,300,350,400,450,500,550,600,650,700,750,800]
 
-        # or a single point
+        #or a single point
         design_ranges = [500]
         specific_energies = [450]
 
         write_logs = False
         if write_logs:
             logging.basicConfig(filename='opt.log', filemode='w', format='%(name)s - %(levelname)s - %(message)s')
-        last_successful_opt = None
-
         # run a sweep of cases at various specific energies and ranges
-        for this_spec_energy in specific_energies:
-            for design_range in design_ranges:
+        for design_range in design_ranges:
+            for this_spec_energy in specific_energies:
                 try:
                     prob = configure_problem()
                     spec_energy = this_spec_energy
@@ -339,13 +295,8 @@ if __name__ == "__main__":
                         prob.model.add_design_var('cruise.hybridization', lower=0.001, upper=0.999)
                         prob.model.add_design_var('climb.hybridization', lower=0.001, upper=0.999)
                         prob.model.add_design_var('descent.hybridization', lower=0.01, upper=1.0)
-                        prob.model.add_design_var('ac|propulsion|thermal|duct|area_nozzle', lower=1., upper=200.)
-                        prob.model.add_design_var('ac|propulsion|thermal|hx|n_wide_cold', lower=100., upper=1500.)
+
                         prob.model.add_constraint('margins.MTOW_margin',lower=0.0)
-                        prob.model.add_design_var('ac|propulsion|thermal|hx|coolant_mass',lower=5.0,upper=15.0)
-
-                        # prob.model.add_constraint('design_mission.residuals.fuel_capacity_margin',lower=0.0)
-
                         prob.model.add_constraint('rotate.range_final',upper=1357)
                         prob.model.add_constraint('v0v1.Vstall_eas',upper=42.0)
                         prob.model.add_constraint('descent.propmodel.batt1.SOC_final',lower=0.0)
@@ -361,11 +312,6 @@ if __name__ == "__main__":
                         prob.model.add_constraint('descent.propmodel.batt1.component_sizing_margin',upper=1.0*np.ones(num_nodes))
                         prob.model.add_constraint('v0v1.propmodel.batt1.component_sizing_margin',upper=1.0*np.ones(num_nodes))
                         prob.model.add_constraint('engineoutclimb.gamma',lower=0.02)
-                        prob.model.add_constraint('climb.propmodel.motorheatsink.T',upper=363.*np.ones(num_nodes))
-                        prob.model.add_constraint('cruise.propmodel.motorheatsink.T',upper=363.*np.ones(num_nodes))
-                        prob.model.add_constraint('climb.propmodel.batteryheatsink.T',upper=323.*np.ones(num_nodes))
-                        prob.model.add_constraint('cruise.propmodel.batteryheatsink.T',upper=323.*np.ones(num_nodes))
-                        prob.model.add_constraint('climb.hxadder.area_constraint',lower=0.001)
                         prob.model.add_objective('mixed_objective') # TODO add this objective
 
                     elif run_type == 'comp_sizing':
@@ -398,10 +344,6 @@ if __name__ == "__main__":
                     if write_logs:
                         filename_to_save = 'case_'+str(spec_energy)+'_'+str(design_range)+'.sql'
                         if os.path.isfile(filename_to_save):
-                            if design_range != 300:
-                                last_successful_opt = filename_to_save
-                            else:
-                                last_successful_opt = 'case_'+str(spec_energy+50)+'_'+str(700)+'.sql'
                             print('Skipping '+filename_to_save)
                             continue
                         recorder = SqliteRecorder(filename_to_save)
@@ -414,31 +356,19 @@ if __name__ == "__main__":
                     prob.setup(check=False)
                     set_values(prob, num_nodes, design_range, spec_energy)
 
-                    if last_successful_opt is not None:
-                        cr = CaseReader(last_successful_opt)
-                        driver_cases = cr.list_cases('driver')
-                        case = cr.get_case(driver_cases[-1])
-                        design_vars = case.get_design_vars()
-                        for key in design_vars.keys():
-                            prob.set_val(key,design_vars[key])
-
                     run_flag = prob.run_driver()
                     if run_flag:
                         raise ValueError('Opt failed')
-                    else:
-                        if write_logs:
-                            last_successful_opt = filename_to_save
-                        prob.cleanup()
+
                 except BaseException as e:
                     if write_logs:
                         logging.error('Optimization '+filename_to_save+' failed because '+repr(e))
                     prob.cleanup()
                     try:
-                        if write_logs:
-                            os.rename(filename_to_save, filename_to_save.split('.sql')[0]+'_failed.sql')
+                        os.rename(filename_to_save, filename_to_save.split('.sql')[0]+'_failed.sql')
                     except WindowsError as we:
                         if write_logs:
                             logging.error('Error renaming file: '+repr(we))
-                            os.remove(filename_to_save)
+                        os.remove(filename_to_save)
 
         show_outputs(prob)
