+++ openconcept/examples/N3_HybridSingleAisle_Refrig.py
@@ -1,15 +1,12 @@
 from __future__ import division
-import sys
-import os
 import numpy as np
 
-sys.path.insert(0, os.getcwd())
 import openmdao.api as om
 import openconcept.api as oc
 # imports for the airplane model itself
 from openconcept.analysis.aerodynamics import PolarDrag
-from examples.aircraft_data.HybridSingleAisle import data as acdata
-from examples.aircraft_data.HybridSingleAisle import MotorFaultProtection
+from openconcept.examples.aircraft_data.HybridSingleAisle import data as acdata
+from openconcept.examples.aircraft_data.HybridSingleAisle import MotorFaultProtection
 from openconcept.analysis.performance.mission_profiles import MissionWithReserve, BasicMission
 from openconcept.components.N3 import N3Hybrid
 from openconcept.components.motor import SimpleMotor
