+++ openconcept/examples/aircraft_data/HybridSingleAisle.py
