+++ openconcept/examples/methods/weights_turboprop.py
