+++ openconcept/examples/propulsion_layouts/simple_all_electric.py
