+++ openconcept/examples/propulsion_layouts/simple_series_hybrid.py
