+++ openconcept/examples/propulsion_layouts/simple_turboprop.py
