+++ openconcept/examples/propulsion_layouts/thermal_series_hybrid.py
